+++ Source/Experiments/MJDTestCryoVac/ORMJDTestCryostats/ORMJDTestCryostat.h
@@ -29,7 +29,7 @@
 {
 	id					 delegate;
 	int					 connectionStatus; 
-	int			         tag;
+	NSUInteger	         tag;
 	NSMutableDictionary* partDictionary;
 	NSMutableDictionary* valueDictionary;
 	NSMutableDictionary* statusDictionary;
@@ -41,7 +41,7 @@
 - (int) connectionStatus;
 - (void) setConnectionStatus:(int) aState;
 - (NSUInteger) tag;
-- (void) setTag:(int)aValue;
+- (void) setTag:(NSUInteger)aValue;
 - (BOOL) showGrid;
 - (void) setDelegate:(id)aDelegate;
 - (void) makeParts;
